Alex dev

+++ FieryLedLampMultilingual/FieryLedLampMultilingual/OtaManager.h
@@ -118,8 +118,8 @@ class OtaManager
     }
 
   private:
-    uint64_t momentOfFirstConfirmation = 0;                 // момент времени, когда получено первое подтверждение и с которого начинается отсчёт ожидания второго подтверждения
-    uint64_t momentOfOtaStart = 0;                          // момент времени, когда развёрнута WiFi точка доступа для обновления по воздуху
+    uint32_t momentOfFirstConfirmation = 0;                 // момент времени, когда получено первое подтверждение и с которого начинается отсчёт ожидания второго подтверждения
+    uint32_t momentOfOtaStart = 0;                          // момент времени, когда развёрнута WiFi точка доступа для обновления по воздуху
     ShowWarningDelegate showWarningDelegate;
 
     void startOtaUpdate()
